v 1.7.1

+++ nightgames/status/Alluring.java
@@ -65,7 +65,7 @@ public class Alluring extends Status {
 	}
 
 	@Override
-	public int pleasure(Combat c, int x) {
+	public double pleasure(Combat c, double x) {
 		return 0;
 	}
