+++ nightgames/status/Charmed.java
@@ -62,7 +62,7 @@ public class Charmed extends Status {
 	}
 
 	@Override
-	public int pleasure(Combat c, int x) {
+	public double pleasure(Combat c, double x) {
 		return 0;
 	}
