+++ nightgames/status/Cynical.java
@@ -52,7 +52,7 @@ public class Cynical extends Status {
 	}
 
 	@Override
-	public int pleasure(Combat c, int x) {
+	public double pleasure(Combat c, double x) {
 		return 0;
 	}
