+++ nightgames/status/Oiled.java
@@ -39,7 +39,7 @@ public class Oiled extends Status {
 	}
 
 	@Override
-	public int pleasure(Combat c, int x) {
+	public double pleasure(Combat c, double x) {
 		return x/4;
 	}
