+++ nightgames/status/Trance.java
@@ -63,7 +63,7 @@ public class Trance extends Status {
 	}
 
 	@Override
-	public int pleasure(Combat c, int x) {
+	public double pleasure(Combat c, double x) {
 		return 0;
 	}
